Separate actions and a common socket service
