+++ materia/src/services/io.js
@@ -1,0 +1,5 @@
+'use strict';
+
+import IO from 'socket.io-client';
+
+export const socket = IO();
